move neo4j to its own file
